fix(tools): robustness & consistency cleanup (issue #62 group 1)
- test-runner: treat parse/schema failure of a test file as a hard
  failure (increment fileErrors; exit 1 if fileErrors > 0) rather than
  silently returning and reporting exit 0
- parity-check: wrap unguarded JSON.parse in try/catch so a malformed
  test file exits 1 cleanly instead of dumping a stack trace
- compile-lists: use Object.create(null) for the entries accumulator to
  avoid the __proto__ silent-data-loss path on a null-proto-keyed entry
- validator: NFC-normalize before lowercasing in the sort check, matching
  the canonical casefold key derivation (spec S3.2); replace localeCompare
  with ordinal comparison to match compile-lists and keep artifacts
  reproducible across locales
- build-lists: standardise sort to ordinal (plain .sort()) to match
  compile-lists; no reordering observed on current ASCII-only data
- extract shared parseCSV (RFC 4180 quote-aware) and casefoldKey into
  tools/build-lists/lib/util.js; import in build-lists, fda-crosscheck,
  and nz-crosscheck -- the NZ crosscheck was previously quote-unaware
  and all three crosschecks were using bare .toLowerCase() instead of
  NFC+toLowerCase for casefold keys
